pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptly"
version = "0.1.0"
description = "A lightweight library for LLM prompt management, observability/tracing, and optimization"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Tucker Leach", email = "leachtucker@gmail.com"},
]
keywords = ["llm", "prompt", "ai", "machine-learning", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "jinja2>=3.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-genai>=0.2.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
cli = [
    "click>=8.0.0",
    "rich>=13.0.0",
]
ui = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "streamlit>=1.20.0",
]

[project.scripts]
promptly = "promptly.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["promptly*"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.format]
# Use double quotes for strings (like Black)
quote-style = "double"
# Indent with spaces (like Black)
indent-style = "space"
# Respect magic trailing comma (like Black)
skip-magic-trailing-comma = false
# Use Unix line endings
line-ending = "auto"